requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.9
